requirements.txt
numpy
imutils
sklearn
tensorflow
keras
opencv-python
selenium
Pillow